Refactor emitter out of compiler.
Precursor to #540
