+++ notebooks/mnist.py
@@ -21,7 +21,7 @@ from torch import nn, optim
 from torch.utils import data as dutils
 from torchrl.data import tensor_specs as tspecs
 
-from aioway.compilers import ClfLogitHead, linear_regression
+from aioway.emits import ClfLogitHead, linear_regression
 from aioway.dsets import DatasetIdxDset, Dset, InputTarget, MnistDataset
 from aioway.trainers import StaticTrainer, TrainCfg
 from aioway.tspecs import as_tspec
