+++ src/aioway/__init__.py
@@ -10,7 +10,8 @@ from . import (
     _utils,
     api,
     app,
-    compilers,
+    emits,
+    rewrites,
     dsets,
     errors,
     instrs,
