--- src/aioway/compilers/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
-
-from .emits import *
-from .rewrites import *
