+++ src/aioway/emits/compound.py
