+++ src/aioway/emits/linear.py
