+++ src/aioway/emits/losses/contrastive.py
